requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
console==0.9907
dnspython==2.2.1
exceptiongroup==1.2.1
ezenv==0.92
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
h11==0.14.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.0
MarkupSafe==2.1.1
months==2.0.0
names==0.3.0
nanoid==2.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
PyJWT==2.4.0
pymongo==4.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
random-utilities==1.0.14
requests==2.28.0
scikit-learn==1.2.2
scipy==1.13.0
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starkbank-ecdsa==2.2.0
threadpoolctl==3.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==1.26.9
Werkzeug==2.1.2
wsproto==1.2.0
zipp==3.8.0
